Feature/cleanup

+++ src/main/java/com/bank/bean/customeraccount/CustomerAccountId.java
@@ -5,12 +5,12 @@ import java.io.Serializable;
 /**
  * Class used for the many-to-many relation on customer-account.
  */
-public class CustomerAccountId implements Serializable{
+public class CustomerAccountId implements Serializable {
     private int customerId;
     private int accountId;
 
     public int hashCode() {
-        return (int)(customerId + accountId);
+        return (customerId + accountId);
     }
 
     public boolean equals(Object object) {
